+++ packages/state/src/index.ts
@@ -1,13 +1,16 @@
 export {
   forcePersistValues,
   loadPersistedValues,
+  persistStoreIntoKey,
+  loadStoreFromKey,
   persist,
+  storeName,
   type PersistOptions,
   type PropertyTransform,
 } from './decorators/persist';
 export { property } from './decorators/property';
 export { isOptionList, isSlider, type Option, type OptionList, type OptionType, type Slider } from './form/option';
-export { getIndex, getOption, nextOption, previousOption } from './form/option-list';
+export { getIndex, getOption, setOption, nextOption, previousOption } from './form/option-list';
 export { decrementSlider, incrementSlider } from './form/slider';
 export { State } from './state';
 export { fileToString, stringToFile } from './transformer/file';